Renamed binaries and config files to something more sensisble.
