+++ src/simple_uam/tools/client/__main__.py
