+++ src/simple_uam/util/config/__init__.py
@@ -4,6 +4,7 @@ from .manager import Config
 from .win_setup_config import WinSetupConfig
 # from .lin_setup_config import LinSetupConfig # noqa
 from .craidl_config import CraidlConfig
+from .broker_config import BrokerConfig
 from .d2c_workspace_config import D2CWorkspaceConfig
 from .d2c_worker_config import D2CWorkerConfig
 from omegaconf import SI, II, OmegaConf
@@ -20,6 +21,7 @@ __all__: List[str] = [
     'WinSetupConfig',
     # 'LinSetupConfig', # No need for this yet
     'CraidlConfig',
+    'BrokerConfig',
     'D2CWorkspaceConfig',
     'D2CWorkerConfig',
 ]  # noqa: WPS410 (the only __variable__ we use)
